+++ mitra_models/src/profiles/types.rs
@@ -58,6 +58,7 @@ pub enum MentionPolicy {
     #[default]
     None,
     OnlyKnown,
+    OnlyConnected,
 }
 
 impl From<&MentionPolicy> for i16 {
@@ -65,6 +66,7 @@ impl From<&MentionPolicy> for i16 {
         match value {
             MentionPolicy::None => 0,
             MentionPolicy::OnlyKnown => 1,
+            MentionPolicy::OnlyConnected => 2,
         }
     }
 }
@@ -76,6 +78,7 @@ impl TryFrom<i16> for MentionPolicy {
         let policy = match value {
             0 => Self::None,
             1 => Self::OnlyKnown,
+            2 => Self::OnlyConnected,
             _ => return Err(DatabaseTypeError),
         };
         Ok(policy)