+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Added documentation for `update-config` command.
 - Generate Ed25519 key for instance actor.
+- Added "Only Connected" mention policy.
 
 ### Changed